requirements.txt
lief==0.12.1
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
simpy==4.0.1
